Update and delete is now working

+++ src/main/java/net/lerraj/spring/dao/StudentsDao.java
@@ -8,7 +8,7 @@ public interface StudentsDao {
 	
 	public void saveOrUpdate(Students student);
 	
-	public void delete(int studentId);
+	public void delete(long studentId);
 	
 	public Students get(int studentId);
